feat(partners): consolidar dashboard en partners module y refactor del portal de propiedades
- Mueve la lógica del dashboard desde el módulo dashboard/ al módulo partners/ (partners.service, partners.controller)
- Elimina el módulo dashboard/ completo (controller, service, types, specs)
- Actualiza app.module.ts y app.controller.ts para reflejar la nueva estructura
- Refactoriza MetricCard, MonthlyChart y ui.tsx del dashboard de organización
- Actualiza el dashboard de organización (index.tsx) y sus tests
- Agrega claves i18n en en.json y es.json para los nuevos controles del dashboard de propiedad
- Mueve dashboard.types.ts a partners/ y amplía los tipos existentes
- Actualiza members module (controller, service, dto, specs) con ajustes de la iteración actual
- Refactoriza la página de propiedad de búsqueda (frontend/src/pages/property/index.tsx)
