+++ frontend/src/i18n/locales/en.json
@@ -470,6 +470,8 @@
       "view_detail": "View detail",
       "no_properties": "No properties registered.",
       "back": "← My hotels",
+      "label": "Property",
+      "edit_property": "Edit property",
       "load_error": "Failed to load properties. Please try again later.",
       "login_required": "Sign in as a partner to view your properties."
     },
