--- services/partners-service/src/dashboard/dashboard.module.ts
@@ -1,11 +1,0 @@
-import { Module } from "@nestjs/common";
-import { ClientsModule } from "../clients/clients.module.js";
-import { DashboardController } from "./dashboard.controller.js";
-import { DashboardService } from "./dashboard.service.js";
-
-@Module({
-  imports: [ClientsModule],
-  controllers: [DashboardController],
-  providers: [DashboardService],
-})
-export class DashboardModule {}
