+++ services/partners-service/src/partners/dashboard.controller.spec.ts
@@ -1,78 +1,47 @@
 import { Test, TestingModule } from "@nestjs/testing";
-import { DashboardController } from "./dashboard.controller.js";
-import { DashboardService } from "./dashboard.service.js";
+import { PartnersController } from "./partners.controller.js";
+import { PartnersService } from "./partners.service.js";
 
-describe("DashboardController", () => {
-  let controller: DashboardController;
-  let getHotelState: jest.Mock;
+describe("PartnersController (dashboard)", () => {
+  let controller: PartnersController;
+  let getPartnerMetrics: jest.Mock;
   let getPayments: jest.Mock;
-  let getProperties: jest.Mock;
 
   beforeEach(async () => {
-    getHotelState = jest.fn().mockResolvedValue("hotel-state");
+    getPartnerMetrics = jest.fn().mockResolvedValue("partner-metrics");
     getPayments = jest.fn().mockResolvedValue("payments");
-    getProperties = jest.fn().mockResolvedValue("properties");
     const moduleRef: TestingModule = await Test.createTestingModule({
-      controllers: [DashboardController],
+      controllers: [PartnersController],
       providers: [
         {
-          provide: DashboardService,
-          useValue: { getHotelState, getPayments, getProperties },
+          provide: PartnersService,
+          useValue: { getPartnerMetrics, getPayments },
         },
       ],
     }).compile();
-    controller = moduleRef.get(DashboardController);
+    controller = moduleRef.get(PartnersController);
   });
 
-  describe("properties", () => {
-    it("delegates to getProperties with partnerId", async () => {
-      await controller.properties("p1");
-      expect(getProperties).toHaveBeenCalledWith("p1");
-    });
-  });
-
-  describe("hotelState", () => {
-    it("uses provided month + roomType, passes null propertyId", async () => {
-      await controller.hotelState("p1", "2026-03", "Doble Superior");
-      expect(getHotelState).toHaveBeenCalledWith(
-        "p1",
-        "2026-03",
-        "Doble Superior",
-        null,
-      );
+  describe("partnerMetrics", () => {
+    it("delegates with partnerId, month, and roomType", async () => {
+      await controller.partnerMetrics("p1", "2026-03", "Suite");
+      expect(getPartnerMetrics).toHaveBeenCalledWith("p1", "2026-03", "Suite");
     });
 
     it("falls back to current month when month is invalid", async () => {
-      await controller.hotelState("p1", "not-a-month", undefined);
-      const call = getHotelState.mock.calls[0] as [
+      await controller.partnerMetrics("p1", "not-a-month", undefined);
+      const call = getPartnerMetrics.mock.calls[0] as [
         string,
         string,
-        string | null,
         string | null,
       ];
       expect(call[1]).toMatch(/^\d{4}-\d{2}$/);
       expect(call[2]).toBeNull();
-      expect(call[3]).toBeNull();
     });
 
     it("treats blank roomType as null", async () => {
-      await controller.hotelState("p1", "2026-03", "   ");
-      expect(getHotelState).toHaveBeenCalledWith("p1", "2026-03", null, null);
-    });
-
-    it("passes propertyId when provided", async () => {
-      await controller.hotelState("p1", "2026-03", undefined, "prop-abc");
-      expect(getHotelState).toHaveBeenCalledWith(
-        "p1",
-        "2026-03",
-        null,
-        "prop-abc",
-      );
-    });
-
-    it("treats blank propertyId as null", async () => {
-      await controller.hotelState("p1", "2026-03", undefined, "   ");
-      expect(getHotelState).toHaveBeenCalledWith("p1", "2026-03", null, null);
+      await controller.partnerMetrics("p1", "2026-03", "   ");
+      expect(getPartnerMetrics).toHaveBeenCalledWith("p1", "2026-03", null);
     });
   });
